Extracted tests to test folder
