+++ package.json
@@ -8,7 +8,7 @@
     "start": "node src/server.js",
     "test": "jest",
     "test-watch": "nodemon --exec \"npm test\"",
-    "travis-test": "jest src/validator.test.js"
+    "travis-test": "jest test/validator.test.js"
   },
   "engines": {
     "node": "9.7.1"
